backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
websockets==13.0
pillow==10.4.0
numpy==1.26.4
python-multipart==0.0.9
